Fix updating of questions within group. 	modified:   app/controllers/question_groups_controller.rb - only remove the original_question 	if it exists and was replaced with a new question.
	modified:   app/views/question_groups/_group_inline_field.html.erb - add question.id
	so questions can be updated.

	modified:   lib/surveyor_gui/models/question_methods.rb - force the questions record to be dirty
	when answers_textbox changes by changing the updated_at attribute.

+++ app/controllers/question_groups_controller.rb
@@ -33,7 +33,7 @@ class QuestionGroupsController < ApplicationController
     if @question_group.save
       #@question_group.questions.update_attributes(survey_section_id: question_group_params[])
       original_question = Question.find(question_group_params[:question_id]) if !question_group_params[:question_id].blank?
-      original_question.destroy
+      original_question.destroy if original_question
       render :inline => '<div id="cboxQuestionId">'+@question_group.questions.first.id.to_s+'</div>', :layout=>'colorbox'
     else
       @title = "Add Question"
